requirements.txt
adium-theme-ubuntu==0.3.4
beautifulsoup4==4.4.1
chardet==2.3.0
Django==1.11.7
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
gyp==0.1
html5lib==0.999
lxml==3.5.0
meld==3.14.2
numpy==1.11.0
Pillow==3.1.2
pygobject==3.20.0
PyJWT==1.5.3
pytz==2017.3
reportlab==3.3.0
scour==0.32
six==1.10.0
uniconvertor==1.1.5
unity-lens-photos==1.0
vboxapi==1.0
